+++ contributing.md
@@ -1,10 +1,10 @@
-Thank you for contributing to React-View-Model!
+Thank you for contributing to ylem!
 
-If you need any help setting up a CanJS development environment and fixing React-View-Model bugs, please reach out to us on the [canjs/canjs Gitter channel](https://gitter.im/canjs/canjs) or email (contact@bitovi.com).  We will happily walk you through setting up a the environment, creating a test, and submitting a pull request.
+If you need any help setting up a CanJS development environment and fixing ylem bugs, please reach out to us on the [canjs/canjs Gitter channel](https://gitter.im/canjs/canjs) or email (contact@bitovi.com).  We will happily walk you through setting up a the environment, creating a test, and submitting a pull request.
 
 ## Reporting Bugs
 
-To report a bug, please visit [GitHub Issues](/canjs/react-view-model/issues).
+To report a bug, please visit [GitHub Issues](/bitovi/ylem/issues).
 
 When filing a bug, it is helpful to include small examples using tools like [JSBin][1] or [CodePen][2].
 
@@ -14,14 +14,14 @@ Search for previous tickets, if there is one add to that one rather than creatin
 
 When contributing, please include tests with new features or bug fixes in a feature branch until you're ready to submit the code for consideration; then push to the fork, and submit a pull request. More detailed steps are as follows:
 
-1. Navigate to your clone of the React-View-Model repository - `cd /path/to/react-view-model`
+1. Navigate to your clone of the ylem repository - `cd /path/to/ylem`
 2. Create a new feature branch - `git checkout -b <branch name>`
 3. Make some changes
 4. Update tests to accomodate your changes
 5. Run tests and make sure they pass in all browsers
 6. Update documentation if necessary
 7. Push your changes to your remote branch - `git push origin <branch name>`
-8. Submit a pull request! Navigate to [Pull Requests](/canjs/react-view-model/pulls) and click the 'New Pull Request' button. Fill in some details about your potential patch including a meaningful title. When finished, press "Send pull request". The team will be notified about your submission and let you know of any problems or targeted release date.
+8. Submit a pull request! Navigate to [Pull Requests](/bitovi/ylem/pulls) and click the 'New Pull Request' button. Fill in some details about your potential patch including a meaningful title. When finished, press "Send pull request". The team will be notified about your submission and let you know of any problems or targeted release date.
 
 [1]: https://jsbin.com/
 [2]: https://codepen.io/
