+++ docs/apis.json
@@ -1,8 +1,8 @@
 [
 	{
-		"react-view-model": [
-			"react-view-model",
-			"react-view-model.Component"
+		"ylem": [
+			"ylem",
+			"ylem.Component"
 		]
 	}
 ]
