+++ docs/component.md
@@ -1,7 +1,7 @@
-@function react-view-model/component react-view-model/component
+@function ylem/component ylem/component
 @parent can-views
 @collection can-ecosystem
-@group react-view-model/component.static 0 static
+@group ylem/component.static 0 static
 
 @description Create an auto-rendering container component with an observable view-model.
 
@@ -10,7 +10,7 @@
 Create an auto-rendering [container component](https://medium.com/@dan_abramov/smart-and-dumb-components-7ca2f9a7c7d0#.v9i90qbq8) by connecting an observable [can-define/map/map] view-model to a React [presentational component](https://medium.com/@dan_abramov/smart-and-dumb-components-7ca2f9a7c7d0#.v9i90qbq8).
 
 ```javascript
-import { Component } from 'react-view-model';
+import { Component } from 'ylem';
 
 export default class AppComponent extends Component {
   ...
@@ -32,22 +32,22 @@ _Note: If you extend any of the [React lifecycle methods](https://facebook.githu
 
 An example application using the ViewModel to create an extra prop, whose value is derived from other props.
 
-@demo demos/react-view-model/component.name.html
+@demo demos/ylem/component.name.html
 
 An example application which includes viewModel mutation and demonstrates auto-rendering.
 
-@demo demos/react-view-model/component.counter.html
+@demo demos/ylem/component.counter.html
 
 You can also play with the above example on JS Bin:
 
-<a class="jsbin-embed" href="https://jsbin.com/lunajov/1/embed?js,output">react-view-model/component demo on jsbin.com</a>
+<a class="jsbin-embed" href="https://jsbin.com/lunajov/1/embed?js,output">ylem/component demo on jsbin.com</a>
 
 You can also use this module with [Preact](https://preactjs.com):
 
-<a class="jsbin-embed" href="https://jsbin.com/fuxerik/2/embed?js,output">react-view-model/component demo with Preact on jsbin.com</a>
+<a class="jsbin-embed" href="https://jsbin.com/fuxerik/2/embed?js,output">ylem/component demo with Preact on jsbin.com</a>
 
 Here’s a recreation of the clock example from [React’s State and Lifecycle docs](https://facebook.github.io/react/docs/state-and-lifecycle.html):
 
-<a class="jsbin-embed" href="https://jsbin.com/zikaxuy/4/embed?js,output">react-view-model/component clock demo on jsbin.com</a>
+<a class="jsbin-embed" href="https://jsbin.com/zikaxuy/4/embed?js,output">ylem/component clock demo on jsbin.com</a>
 
 <script src="https://static.jsbin.com/js/embed.min.js?4.0.4"></script>
