+++ docs/component.name.html
@@ -1,7 +1,7 @@
 <script src="../../node_modules/steal/steal.js" main="@empty" id="demo-source">
 import React from "react";
 import ReactDOM from "react-dom";
-import Component from "react-view-model/component";
+import Component from "ylem/component";
 import DefineMap from "can-define/map/map";
 
 class AppComponent extends Component {
