+++ docs/function.counter.html
@@ -2,7 +2,7 @@
 var React = require("react");
 var ReactDOM = require("react-dom");
 var DefineMap = require("can-define/map/map");
-var reactViewModel = require("react-view-model");
+var reactViewModel = require("ylem");
 
 var ViewModel = DefineMap.extend("AppVM", {
   count: {
